Add server code for Packet Train test
