+++ api/packet.go
@@ -1,8 +1,17 @@
 package api
 
+import "time"
+
 // Packet represents the packet sent for network testing.
 type Packet struct {
 	Sequence  int    // Sequence number.
 	Timestamp int64  // Timestamp (sent).
 	Data      []byte // Data transmitted.
+}
+
+// Received represents the structure received over the network.
+type Received struct {
+	*Packet
+	Received time.Time
+	Size     int64
 }
