+++ go.mod
@@ -4,9 +4,13 @@ go 1.22
 
 require github.com/m-lab/go v0.1.72
 
-require github.com/stretchr/testify v1.9.0 // indirect
+require (
+	github.com/pkg/errors v0.8.1 // indirect
+	github.com/stretchr/testify v1.9.0 // indirect
+)
 
 require (
+	github.com/apex/log v1.9.0
 	github.com/sirupsen/logrus v1.9.3
 	golang.org/x/sys v0.19.0 // indirect
 )
