+++ static/spec.go
@@ -8,4 +8,7 @@ const (
 	PacketBytes = 726
 	PairCount   = 30
 	PairDelay   = 10000 * time.Microsecond
+	TrainCount  = 500
+	TrainDelay  = 1 * time.Second
+	TrainLength = 30
 )
